fix laplace probabilities
originally assumed we had all counts, not just counts of n-grams we have
seen. putting all counts in memory (in ngram.py) was too much, so this
seemed to be the better approach. the basic idea is to compute the
laplace probabilities for the items we have seen in
_generate_probabilities, and then in get_probabilities, if a
probability ever is zero (which it should not be, by laplace
definition), then return what the laplace probability should have been
set to